Do not download a multi-ep torrent if we don't need all the episodes in the season.
Note that we have no way of knowing the torrent's contents, so we cannot
selectively download the subset of files from the multi-ep torrent that we
actually need. We also make the assumption here that a multi-ep torrents
actually contains all the season episodes, so we only download it if we
need the full season, while the torrent might actually contain only a
subset of the season episodes.